fix a bug that always reformat disk

+++ deploy/image/main/scripts/setup-disk.sh
@@ -4,10 +4,17 @@ set -xeu
 device=$1
 mount_point=$2
 
-# setup data disk (only if device is a empty block)
-mount | grep $device -q || {
-    mkfs.ext4 $device
-    echo "$device $mount_point ext4 defaults 0 2" >> /etc/fstab
-    mkdir $mount_point
-    mount $mount_point
+blkid $device || {
+    if [ $? = 2 ]; then
+        # got a blank block
+        mkfs.ext4 $device
+    else
+        exit $?
+    fi
 }
+
+UUID=$(blkid -o value $device | head -n 1)
+echo "UUID=$UUID $mount_point ext4 defaults 0 2" >> /etc/fstab
+
+mkdir $mount_point
+mount $mount_point
